requirements.txt
openai==0.28.0
pillow==10.2.0
bcrypt==4.1.2
